NOMERGE: Place TLV size into TLV itself
To allow the signatures to be replaced, move the size of the TLV into a
small "info" header at the start of the TLV.

Unfortunately, this causes the image swapping to loose robustness
against untimely powerdowns.  In the middle of a swap, the sectors of
the two images are mixed up, and it is no longer valid to try reading
the TLV info header.  Some possible ideas to fix this:

  - Don't use the copy_size stuff, and just copy the full partitions (up
    until the trailer.
  - Compute the image size once at the start, and add a field to the
    trailer to hold it.  On subsequent boots, use the information in the
    trailer.
  - Try to figure it out in the middle.  Given that the TLV has no
    alignment requirements, it might even be split across the two images.
    This doesn't seem reasonable, and is excessively complex.

Based on work by Marko Kiiskila <marko@runtime.io>

Signed-off-by: David Brown <david.brown@linaro.org>