sim: Add 'checked-read' feature to the simulator
Some devices are starting to use ECC to validate the integrity of data
written to flash.  A consequence of this is that reads from data that
was either partially written, or just erased an unwritten will cause the
flash driver's read call to return an error instead of just reading back
unwritten data.  Add support for this to the simulator, using a new
'checked-read' feature.  When this is enabled any reads from unwritten
data will return `-EIO` instead of reading data.

Signed-off-by: David Brown <david.brown@linaro.org>